+++ R/listdial.R
@@ -78,7 +78,6 @@ dial <- function(socket,
   } else {
     .Call(rnng_dialer_create, socket, url)
   }
-
   is.integer(res) && return(invisible(res))
 
   if (is.environment(socket)) {
@@ -182,7 +181,6 @@ listen <- function(socket,
   } else {
     .Call(rnng_listener_create, socket, url)
   }
-
   is.integer(res) && return(invisible(res))
 
   if (is.environment(socket)) {